mod:improved HDF5 dumping in overlapcalculator
currently, overlapcalculator dumps everything again and again in every
cycle, which is quite wasteful. by using the internal HDF5 functions and
a proper data_model this situation could be improved.

Right now the CDD strings are not dumped properly. Untils this is fixed,
the improved HDF5 dump will be disabled by default.