add: allow scaling of 2d anapots by a float
by providing, e.g., scale=1e2 to the constructor of 2d analytical
potentials the energies and its derivatives can be scaled (down/up).

currently this is probably only supported by AnaPot, because of the way
I call the superclass constructor in the derived potentials. Right now
additional kwargs are probably ignored in other 2d analytical
potentials. This will be changed in the future, so the "scale" argument
is supported by all calculators.