`align_factor` has been added as an attribute of `Optimizer.py`. All derived classes use this attribute to pass to the `fit_rigid` function in `helpers.py`
Now that each derived class has access to this flag, the static variable for setting the `align_factor` in the `procrustes` function is removed (it was convenient, but definitely bad practice). Also corrected `1` and `0` to be `1.0` and `0.0`. I should have done this initially since it prevents implicit type casting and improves readability. I also improved the readability of `align_factor` in the function and it is always used as suggested.