add: calculation of rotation matrix to align coordiantes along a vector
can be used to rotate coordinates in a way, to make a bond distance
vector parallel to another vector, e.g., the x-axis. This is useful for
calculations with external field. With the present functionality a
certain bond vector could be aligned parallel to the x-axis and then the
electrical field is applied along the x-axis.